Second commit for SCS components
This commit adds the SCS version of "remove animation from Cloak of Mirroring etc", which uses 291 to block the animations on the items rather than the original ancient method of just erasing the animation files.

Pro tem, I've done this as a *replacement* for the existing component, since historically it's pretty clear that removing the animations from short-duration spells was collateral damage rather than the actual intent. But this is a judgement call; clearly there's also a case for coding it as a subcomponent, but that should be easy enough to implement if you prefer it.