1 gp items for various tweaks
The Two-Handed weapon components (katanas, bastard swords, axes) and the PnP protection items component add an ability
to create items. Unfortunately the ability lacked the 'recharge after resting' flag so that the items could not be
sold for more than 1 gp if discharged and, despite the claims of the readme, resting would not help.