Fixes for Make Khalid a Fighter-Mage
This was causing issues when revisiting the FAI if you didn't recruit Khalid the first time around. It had an unnecessarily complex spawn script which, given that we're outright replacing Khalid's original creature files, is also wholly redundant and removed. Gone, too, is the BGT-specific d file as it's handled by the CPM libraries.

Otherwise, it's the standard 'modernize this, streamline that' bits and bobs. Most notably, there's now a khalid7 for SoD games.