Update unique_containers_fixes.tpa
The bhxspell.sto does not exist anymore in the updated Bone Hill Mod for EE.
The itm and relate sto were never used by the mod anyway and the scrolls that the fix tries to remove never existed. The whole package was some unfinished idea in the mod.
The non-EE mod is not maintained, thus making the fix conditional should be sufficient.