New component, infinite nonmagical ammo
Based on a suggestion in Discord (and some boredom over the weekend) a new component has been added. This will give all bows, slings, and crossbows the ability to shoot unlimited ammo, e.g. like how the Sling of Everard or Short Bow of Gesen, except that the ammo is just the basic bogstandard ammo for the weapon.